Refactor prepare iter to return a stream
