+++ crates/sparrow-main/src/serve/compute_service.rs
@@ -545,7 +545,6 @@ mod tests {
         )
         .await
         .unwrap()
-        .stream()
         .collect()
         .await;
